First sort of working sketch of the REST/HTTP API
Added very basic entry point in bin/rubinius-webconsole
Added 'rubinius-webconsole' executable to gem and made it the default.
Added 'rubinius-debugger' as a dev dependency.
Added 'sinatra' as runtime dependency.
Added the Agent wrapper class meant to be composed of a real RBX Agent.
Added the AgentManager class meant to serve as a container for Agent
instances.
Added first version of HTTP API using Sinatra in RestApi class.
NO tests, because TDD is dead <3 ;)